# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "UserPost demo API",
        "description": "Node.js and Express.js Application API",
        "license": {
            "name": "Post",
            "url": "https://post.com/"
        }
    },
...
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "in": "cookie",
            "name": "token"
        }
    },
    "security": [
        {
            "jwt": []
        }
    ],
    "paths": {
...
        "/post/addpost":{
          "post":{
            "tags":["Post"],
            "summary":"Create New Post",
            "parameters":[
              {
                "in"          : "formData",
                "name"        : "file",
                "type"        : "file",
                "description" : "file to be upload",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["image"]
                    }
                  ]
                }
              },
              {
                "in"          :  "formData",
                "name"        :  "title",  
                "type"        :  "string",
                "description" :  "title of file content",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["title"]
                    }
                  ]
                }
              },
              {
                "in"          :  "formData",
                "name"        :  "description",  
                "type"        :  "string",
                "description" : "content of file content",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["content"]
                    }
                  ]
                }
              }
            
            ],
            "produces" : ["application/json"],
            "responses" :{
              "200":{
                "description":"Create new Post",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status" : { "type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref":"#/definitions/Post"},
                            {
                              "type" : "object",
                              "properties":{
                                "_id"         : {"type" : "string"}
                              } 
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }

        },
...
        "/post/deletepost/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of Post",
              "type":"string"
            }
          ],
          "get":{
            "tags":["Post"],
            "summary":"Delete Post by ID",
            "responses":{
              "200":{
                "description":"Delete Post by ID",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status"  : { "type" : "boolean"},
                        "message" : { "type" : "string"}
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "/post/updatepost/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of Post",
              "type":"string"
            }
          ],
          "post":{
            "tags":["Post"],
            "summary":"Update Post by ID",
            "parameters":[
              {
                "in"          : "formData",
                "name"        : "file",
                "type"        : "file",
                "description" : "file to be upload",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["image"]
                    }
                  ]
                }
              },
              {
                "in"          :  "formData",
                "name"        :  "title",  
                "type"        :  "string",
                "description" :  "title of file content",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["title"]
                    }
                  ]
                }
              },
              {
                "in"          :  "formData",
                "name"        :  "description",  
                "type"        :  "string",
                "description" : "content of file content",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Post"
                    },
                    {
                      "type" : "object",
                      "required" : ["content"]
                    }
                  ]
                }
              }
            ],
            "produces":["application/json"],
            "responses":{
              "200":{
                "description":"Update Post by ID",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref": "#/definitions/Post"},
                            {
                              "type":"object",
                              "properties":{
                                "_id"  : { "type" : "string"},
                                "user" : { "type" : "string"}
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "/post/addcomment/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of Post",
              "type":"string"
            }
          ],
          "post":{
            "tags":["Post"],
            "summary":"Add Comment for Post",
            "parameters":[
              {
                "in"          : "body",
                "name"        : "comment",
                "type"        : "string",
                "description" : "add comment on post",
                "schema":{
                  "allOf" :[
                    {
                      "$ref":"#/definitions/Comment"
                    },
                    {
                      "type" : "object",
                      "required" : ["comment"]
                    }
                  ]
                }
              }
            ],
            "produces":["application/json"],
            "responses":{
              "200":{
                "description":"Update comment Post by ID",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status":{"type" : "boolean"},
                        "data":{
                          "allof":[
                            {"$ref": "#/definitions/Post"},
                            {
                              "type":"object",
                              "properties":{
                                "_id"  : { "type" : "string"},
                                "user" : { "type" : "string"}
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "/post/delete/{id}":{
          "parameters":[
            {
              "name":"id",
              "in":"path",
              "required":true,
              "description":"ID of PostComment",
              "type":"string"
            }
          ],
          "get":{
            "tags":["Post"],
            "summary":"Delete Comment by ID",
            "responses":{
              "200":{
                "description":"Delete comment by ID",
                "schema":{
                  "allOf":[
                    {
                      "type":"object",
                      "properties":{
                        "status"  : { "type" : "boolean"},
                        "message" : { "type" : "string"}
                      }
                    }
                  ]
                }
              }
            }
          }
        },
...
    },
    "definitions": {
...
      "Post": {
        "properties" : {
          "title"         :{ "type" : "string"},
          "description"   :{ "type" : "string"}
        }
      },
      "Comment": {
        "properties" : {
          "comment"        :{ "type" : "string"}
        }
      }
    }
}